api_service/requirements.txt
# Flask
Flask==3.0.0
Flask-CORS==4.0.0

# Validation
pydantic==2.5.3

# Redis
redis==5.0.1

# ML (load saved models)
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
numpy==1.26.2
pandas==2.1.4

# Testing
pytest==7.4.3
requests==2.31.0
